requirements.txt
GitPython==3.1.27
gitdb==4.0.9
smmap==5.0.0